chore: #1916 refactor billing page
