--- packages/admin-portal/src/components/pages/admin-billing/index.ts
@@ -1,1 +1,0 @@
-export { default } from './admin-billing'
