+++ packages/admin-portal/src/components/pages/billing/__tests__/billing.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { shallow } from 'enzyme'
-import AdminBilling from '../admin-billing'
+import AdminBilling from '../billing'
 
 describe('AdminBilling', () => {
   it('should match a snapshot when no error', () => {
