+++ packages/admin-portal/src/components/pages/billing/index.ts
@@ -1,0 +1,1 @@
+export { default } from './billing'
